\name{noontemp}
\alias{noontemp}
\title{Noon Temperatures in Winnipeg, Manitoba}
\usage{data(noontemp)}
\description{
Time Series of noon temperature observations from the Winnipeg
International Airport from January 1, 1960 through December 31, 1980.
}
\format{
  A single vector.  
}
\keyword{datasets}
