\name{testfun}
\alias{testfun}
\title{Functions for Testing Purposes}
\description{
 Functions that can be used in simulations to test the
     effectiveness of the sharpening procedures.
}
\usage{
 testfun(x, k) 
}
\arguments{
\item{x}{numeric vector}
\item{k}{a numeric constant that controls the height of the peak of
          the test function; if missing, a periodic function is
          supplied
}
}
\value{
numeric vector of function output
}
\author{D.Wang}
\keyword{models}
